Move legacy, add cookie utils
