+++ src/app/Routes.tsx
@@ -8,7 +8,7 @@ import {
 
 import Auth from '../features/auth/Auth';
 import Count from '../features/count/Counter';
-import Legacy from '../features/legacy/LegacyNew';
+import Legacy from '../features/legacy/Legacy';
 import Navigator from '../features/navigator/Navigator';
 import PageNotFound from '../features/layout/PageNotFound';
 import ProfileWrapper from '../features/profile/Profile';
