+++ src/common/cookie.ts
@@ -9,3 +9,19 @@ export function getCookie(key: string) {
     return undefined;
   }
 }
+
+export function setCookie(
+  name: string,
+  value: string,
+  options: { expires?: Date; path?: string } = { path: '/' }
+) {
+  const { expires, path } = options;
+  let cookieString = `${name}=${value}`;
+  if (expires) cookieString += `; expires=${expires.toUTCString()}`;
+  if (path) cookieString += `; path=${path}`;
+  document.cookie = cookieString;
+}
+
+export function clearCookie(name: string) {
+  setCookie(name, '', { expires: new Date('Thu, 01 Jan 1970 00:00:00 GMT') });
+}
